Convert errors to failure crate.
On the library side, wrap errors from our dependencies with an enum, and derive
Fail for it. This approach to lets us introduce strong types with custom
printing fairly quickly. It has the disadvantage of being somewhat rigid, and
is one of a few approaches described here:
https://boats.gitlab.io/failure/custom-fail.html

On the binary side, as a user of the library, failure allows us to make liberal
use of the `?` operator, even in main, since main can do `Result` now.

Refs #1