Create application config module
config::Config can be deserialized from a config file at

~/.config/fitbit-grabber/conf.toml

or set with --config at the command line.

The config struct is all Options that must be unwrapped, so we could have
separate command funcs that care about particular values on a Config.

Also ran rustfmt on the whole project. Refs #1